requirements.txt
defusedxml~=0.7.1
ipython~=9.4.0
PyQt6~=6.9.1
numpy~=2.3.2
pillow~=11.3.0
pip~=25.1.1
attrs~=25.3.0
distro~=1.9.0
wheel~=0.45.1
protobuf~=6.31.1
cryptography~=45.0.5
filelock~=3.18.0
pytest~=8.4.1
hypothesis~=6.136.2
charset-normalizer~=3.4.2
Cython~=3.1.2
setuptools~=80.9.0
pytz~=2025.2
overrides~=7.7.0
cffi~=1.17.1
typing_extensions~=4.14.1
sympy~=1.14.0
mpmath~=1.3.0
gmpy2~=2.2.1
lark~=1.2.2
decorator~=5.2.1
torch~=2.7.1
matplotlib~=3.10.3
onnx~=1.18.0
tqdm~=4.67.1
scipy~=1.16.0
psutil~=7.0.0
packaging~=25.0
colorama~=0.4.6
networkx~=3.5
urllib3~=2.5.0
zstandard~=0.23.0
requests~=2.32.4
ujson~=5.10.0
libarchive-c~=5.1
MarkupSafe~=3.0.2
pandas~=2.3.1
PyYAML~=6.0.2
trove-classifiers~=2025.5.9.12
build~=1.2.2
platformdirs~=4.3.6